requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
django-oauth-toolkit==2.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
jwcrypto==1.3
Markdown==3.3.6
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.1
zipp==3.7.0
